Basic mockup setup
