+++ frontend/src/index.js
@@ -2,46 +2,19 @@ import React from "react";
 import ReactDOM from "react-dom";
 //import './index.css';
 //import App from './App';
-import SearchPage from "./search";
-import ProfessorList from "./professorList.js";
 
-const pagesToUse = [<ProfessorList />, <SearchPage />]
+import App from "./App.js";
+/*
+const pagesToUse = [<ProfessorList />, <SearchPage />, <ProfessorPage/>]
 var pageOn = 0;
 
 
 ReactDOM.render(pagesToUse[pageOn], document.getElementById("root"));
+*/
 
-
-async function getProfessors(){
-   try {
-      //const response = await axios.post('');
-      const response = [
-         new Professor("Bruno Da Silva", "CSC", true),
-         new Professor("Kurt Mammen", "CSC", false),
-         new Professor("Christopher Siu", "CSC", true),
-         new Professor("Farah Al Nakib", "HIST", true)
-     ];
-      return response;
-   }
-   catch (error) {
-      console.log(error);
-      return false;
-   }
-}
-
-async function getProfessor(){
-   try {
-      //const response = await axios.post('');
-      const response = new Professor("Bruno Da Silva", "CSC", true);
-      response.reviews = [new Review("Winter", 2015, "this is a test review", 4.0, "CSC 307"),
-      new Review("Fall", 2018, "... I have more things to sayyyyyy", 1.0, "CSC 321"),
-      new Review("Spring", 2021, "this is another test review", 4.0, "CSC 307"),
-      new Review("Fall", 2021, "BLAGH BLAGH BLAGH", 3.0, "CSC 349")];
-      
-      return response;
-   }
-   catch (error) {
-      console.log(error);
-      return false;
-   }
-}
+ReactDOM.render(
+   <React.StrictMode>
+     <App />
+   </React.StrictMode>,
+   document.getElementById('root')
+ );
